+++ lib/formStateProvider.js
@@ -14,7 +14,7 @@ var React = require("react");
 var PropTypes = require("prop-types");
 var immutable_1 = require("immutable");
 var definitionChecker_1 = require("./definitionChecker");
-var eventAdaptors_1 = require("./eventAdaptors");
+var eventHandlerAdaptors_1 = require("./eventHandlerAdaptors");
 var handlerEngine_1 = require("./handlerEngine");
 /**
  * This HOC enhances your form component.
@@ -140,8 +140,8 @@ function formStateProvider(Form) {
                     formSubmit: this.submit,
                     formReset: this.reset,
                     // FormPropsEx
-                    formOnChange: eventAdaptors_1.changeAdaptor(this.change),
-                    formOnValidate: eventAdaptors_1.focusAdaptor(this.validate),
+                    formOnChange: eventHandlerAdaptors_1.changeAdaptor(this.change),
+                    formOnValidate: eventHandlerAdaptors_1.focusAdaptor(this.validate),
                 }).toJS();
                 return React.createElement(Component, props);
             };
