+++ lib/index.d.ts
@@ -1,4 +1,4 @@
 export * from './types';
 export * from './formStateProvider';
-export * from './eventAdaptors';
-export * from './memorizedAdaptors';
+export * from './eventHandlerAdaptors';
+export * from './validatorAdaptors';
