+++ lib/index.js
@@ -4,6 +4,6 @@ function __export(m) {
 }
 Object.defineProperty(exports, "__esModule", { value: true });
 __export(require("./formStateProvider"));
-__export(require("./eventAdaptors"));
-__export(require("./memorizedAdaptors"));
+__export(require("./eventHandlerAdaptors"));
+__export(require("./validatorAdaptors"));
 //# sourceMappingURL=index.js.map
